Feat: 데이터베이스에 스킬 추가
SkillDB에 실제 스킬을 추가했고, 제대로 작동하지 않던 기능을 수정했습니다.

+++ DungeonTRPG/EntitySystem/ActiveEffect/HpDownWithATK.cs
@@ -11,9 +11,9 @@ namespace DungeonTRPG.EntitySystem.ActiveEffect
     internal class HpDownWithATK: IEffect
     {
         int amount;
-        int modifier;
+        float modifier;
         int TotalDamage;
-        public HpDownWithATK(int amount, int modifier)
+        public HpDownWithATK(int amount, float modifier)
         {
             this.amount = amount;
             this.modifier = modifier;
@@ -23,7 +23,7 @@ namespace DungeonTRPG.EntitySystem.ActiveEffect
         //범용적으로 활용 가능(amount=0, modifier=1.5 / amount=10, modifier=0.5...)
         public void UseEffect(Character caster, Character target)
         {
-            TotalDamage = amount + (caster.Stat.Atk*modifier);
+            TotalDamage = (int)(amount + (caster.Stat.Atk*modifier));
             target.Stat.SetDamage(amount);
         }
     }
